some basic unfinished refactoring
